Rename CompanySummar component
