+++ components/companycontent/CompanySummary.tsx
